third_party/WebKit/LayoutTests/external/wpt/FileAPI/url/url-with-xhr.any-expected.txt

This is a testharness.js-based test.
PASS Blob URLs can be used in XHR
PASS XHR with a fragment should succeed
FAIL XHR of a revoked URL should fail assert_equals: expected 0 but got 404
PASS Only exact matches should revoke URLs, using XHR
FAIL Appending a query string should cause XHR to fail assert_equals: expected 0 but got 404
FAIL Appending a path should cause XHR to fail assert_equals: expected 0 but got 404
PASS XHR with method "HEAD" should fail
PASS XHR with method "POST" should fail
PASS XHR with method "DELETE" should fail
PASS XHR with method "OPTIONS" should fail
PASS XHR with method "PUT" should fail
PASS XHR with method "CUSTOM" should fail
PASS XHR should return Content-Type from Blob
FAIL Revoke blob URL after open(), will fetch assert_equals: expected "test blob contents" but got ""
Harness: the test ran to completion.
